Exception Mapping

+++ src/main/java/org/geekocon/services/TypeService.java
@@ -1,9 +1,14 @@
 package org.geekocon.services;
+import org.geekocon.dto.Zone;
 import org.geekocon.dto.ZoneType;
+import org.geekocon.exception.DependencyZoneTypeException;
 
 import javax.inject.Singleton;
 import javax.transaction.Transactional;
+import javax.ws.rs.core.Response;
 import java.util.List;
+
+import static javax.ws.rs.core.Response.Status.OK;
 
 @Singleton
 public class TypeService {
@@ -19,10 +24,11 @@ public class TypeService {
     }
 
     @Transactional
-    public void deleteType(Long id){
+    public Response deleteType(Long id){
+        Zone temp = Zone.find("type_id = ?1", id).firstResult();
+        if(temp != null) throw new DependencyZoneTypeException(temp.getName());
         ZoneType.deleteById(id);
-       // throw new DependencyZoneTypeException(zone.getName());
-        return;
+        return Response.status(OK).build();
     }
 
 }
